Adds better UI and UX. Adds hover to cards in feed, making it more interactive. Files have been reorganised and an API folder have been made for API calls. Upload section gets a fixed position on top for better scrolling experience
