+++ feed/index.html
@@ -55,7 +55,8 @@
     </div>
 
     <img
-      src="//img.freepik.com/free-psd/3d-icon-social-media-app_23-2150049569.jpg?t=st=1709563197~exp=1709566797~hmac=b689e5e177c3d6c62da590cd067d75d9ba4e97609d314b94de53d37a913bf5a5&w=740"
+      src="https://img.freepik.com/free-psd/3d-icon-social-media-app_23-2150049569.jpg?t=st=1709563197~exp=1709566797~hmac=b689e5e177c3d6c62da590cd067d75d9ba4e97609d314b94de53d37a913bf5a5&w=740"
+      ;
       alt="A default profile avatar"
       class="rounded-circle float-end col-1"
     /><!-- This is Bootstrap icons -->
@@ -81,7 +82,9 @@
             >
           </li>
           <li class="nav-item">
-            <a class="nav-link d-block d-sm-none" href="../index.html"
+            <a
+              class="nav-link d-block d-sm-none my-hover-enlarge"
+              href="../index.html"
               >Log-out</a
             >
           </li>
@@ -96,7 +99,7 @@
       <!-- Upload section -->
 
       <div
-        class="card my-border border-rounded border-3 border-info bg-secondary p-3 mt-5 col-12 col-sm-5"
+        class="card my-border border-rounded border-3 border-info h-25 bg-secondary p-3 mt-5 col-12 col-sm-5 position-sticky top-0"
       >
         <h1 class="lead fw-bold text-center">Hi there!</h1>
         <img
@@ -130,7 +133,7 @@
               id="tags"
             />
             <button
-              class="bolder btn btn-info m-4 float-end"
+              class="bolder btn btn-info mt-3"
               id="postButton"
               type="button"
             >
