+++ feed/single-post/index.html
@@ -92,7 +92,11 @@
             >
           </li>
           <li class="nav-item">
-            <a class="nav-link d-block d-sm-none" href="/index.html">Log-out</a>
+            <a
+              class="nav-link d-block d-sm-none my-hover-enlarge"
+              href="/index.html"
+              >Log-out</a
+            >
           </li>
         </ul>
       </div>
