+++ index.html
@@ -100,7 +100,7 @@
               placeholder="Enter password"
               pattern="(?=.*\d)(?=.*[a-z])(?=.*[A-Z]).{8,}"
               title="For security reasons, your password must be at least 8 characters long and contain at least one uppercase letter, one lowercase letter, and one number."
-              autocomplete="new-password"
+              autocomplete="current-password"
               required
             />
           </div>
@@ -132,7 +132,13 @@
       class="container-fluid border border-info border-5 text-center p-2 mb-0 mt-5"
     >
       &copy2024
-      <div><h5 class="display-6">*musync*</h5></div>
+      <div>
+        <img
+          src="/images/_musync_.png"
+          alt="A logo in bold letters, it spells *musync*"
+          class="col-2"
+        />
+      </div>
     </footer>
   </body>
 </html>
