+++ profile/index.html
@@ -78,7 +78,9 @@
             <a class="nav-link" href="../feed/index.html">Feed</a>
           </li>
           <li class="nav-item">
-            <a class="nav-link d-block d-sm-none" href="../index.html"
+            <a
+              class="nav-link d-block d-sm-none my-hover-enlarge"
+              href="../index.html"
               >Log-out</a
             >
           </li>
