+++ register/index.html
@@ -121,7 +121,13 @@
       class="container-fluid border border-info border-5 text-center p-2 mb-0 mt-5"
     >
       &copy2024
-      <div><h5 class="display-6">*musync*</h5></div>
+      <div>
+        <img
+          src="/images/_musync_.png"
+          alt="A logo in bold letters, it spells *musync*"
+          class="col-2"
+        />
+      </div>
     </footer>
   </body>
 </html>
