+++ src/scss/index.scss
@@ -49,7 +49,7 @@ $border-radius2: 5px;
 }
 
 $form-border: 5px;
-$form-size: 1rem;
+$form-size: 0.6rem;
 $form-border-thickness: 3px;
 $form-type: solid;
 $form-border-color: #3590d2;
@@ -73,3 +73,11 @@ $form-size-min: 0.5rem;
 }
 
 $link-decoration: none;
+
+.my-hover-enlarge {
+  transition: transform 0.2s;
+
+  &:hover {
+    transform: scale(1.1);
+  }
+}
